Updated multinormal.py

+++ math/multivariate_prob/multinormal.py
@@ -31,4 +31,4 @@ class MultiNormal:
         normal_const = 1.0 / np.sqrt((2 * np.pi) ** k * det_cov)
         deviation = x - self.mean
         expon = -0.5 * np.dot(deviation.T, np.dot(inv_cov, deviation)).item()
-        return normal_const * np.exp(expon)
+        return float(normal_const * np.exp(expon))
